fix- correção de tipagens

+++ projeto-tfg/src/main/kotlin/school/sptech/projetotfg/dto/EnderecoDTO.kt
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.Size
 
 data class EnderecoDTO(
     @field:NotBlank val logradouro: String,
-    @field:Positive val numero: Int,
+    @field:Positive val numero: Long,
     @field:Size(min = 8, max = 8) val cep: String,
     @field:NotBlank val bairro: String,
     @field:NotBlank val cidade: String
